Add (currently optional) slot for ui component into the nodeInfoFactory()
