+++ inst/editor/src/ui-node-definitions/Shiny/ShinyActionButton/input_action_button.tsx
@@ -1,5 +1,9 @@
-import type { CSSMeasure } from "../inputFieldTypes";
-import { nodeInfoFactory } from "../nodeInfoFactory";
+import Button from "../../../components/Inputs/Button/Button";
+import buttonIcon from "../../assets/icons/shinyButton.png";
+import type { CSSMeasure } from "../../inputFieldTypes";
+import { nodeInfoFactory } from "../../nodeInfoFactory";
+
+import classes from "./styles.module.css";
 
 export const input_action_button = nodeInfoFactory<{
   inputId: string;
@@ -42,4 +46,13 @@ export const input_action_button = nodeInfoFactory<{
   category: "Inputs",
   description:
     "Creates an action button whose value is initially zero, and increments by one each time it is pressed.",
+  ui_component: ({ namedArgs, wrapperProps }) => {
+    const { label = "My Action Button", width } = namedArgs;
+
+    return (
+      <div className={classes.container} {...wrapperProps}>
+        <Button style={width ? { width } : undefined}>{label}</Button>
+      </div>
+    );
+  },
 });
