+++ inst/editor/src/ui-node-definitions/registered_ui_nodes.ts
@@ -1,79 +1,4 @@
-import { bslibCardBodyInfo, bslibCardHeaderInfo, bslibCardInfo } from "./Bslib";
-import { bslibCardFooterInfo } from "./Bslib/card_footer";
-import { bslibNavPanelInfo } from "./Bslib/NavPanel";
-import { bslibSidebar } from "./Bslib/Sidebar";
-import { bslibValueBoxInfo } from "./Bslib/ValueBox/ValueBox";
-import { dtDTOutputInfo } from "./DT";
-import {
-  gridlayoutCardInfo,
-  gridlayoutGridCardPlotInfo,
-  gridlayoutGridContainerInfo,
-  gridlayoutGridPageInfo,
-  gridlayoutTextPanelInfo,
-} from "./gridlayout";
-import { testingErrorNodeInfo } from "./internal/__TestingErrorNode";
-import { textNodeInfo } from "./internal/TextNode";
-import { unknownUiFunctionInfo } from "./internal/UnknownUiFunction";
-import { plotlyPlotlyOutputInfo } from "./plotly";
-import { shinyActionButtonInfo } from "./Shiny/ShinyActionButton";
-import { shinyCheckboxGroupInputInfo } from "./Shiny/ShinyCheckboxGroupInput";
-import { shinyCheckboxInputInfo } from "./Shiny/ShinyCheckboxInput";
-import {
-  shinyLayoutSidebarInfo,
-  shinyPanelMainInfo,
-} from "./Shiny/ShinyLayoutSidebar/LayoutSidebar";
-import { markdownNodeInfo } from "./Shiny/ShinyMarkdown/markdown";
-import { shinyNavbarPageInfo } from "./Shiny/ShinyNavbarPage";
-import { shinyNumericInputInfo } from "./Shiny/ShinyNumericInput";
-import { shinyPlotOutputInfo } from "./Shiny/ShinyPlotOutput";
-import { shinyRadioButtonsInfo } from "./Shiny/ShinyRadioButtons/ShinyRadioButtons";
-import { shinySelectInputInfo } from "./Shiny/ShinySelectInput";
-import { shinySliderInputInfo } from "./Shiny/ShinySliderInput";
-import { shinyTabsetPanelInfo } from "./Shiny/ShinyTabsetPanel";
-import { shinyTextInputInfo } from "./Shiny/ShinyTextInput";
-import { shinyTextOutputInfo } from "./Shiny/ShinyTextOutput/ShinyTextOutput";
-import { shinyUiOutputInfo } from "./Shiny/ShinyUiOutput";
-
-/**
- * This is the main object that contains the info about a given uiNode. Once the
- * node info object is created and added here the ui-node will be usable within
- * the editor
- */
-export const registered_ui_nodes = [
-  shinyActionButtonInfo,
-  shinyNumericInputInfo,
-  shinySliderInputInfo,
-  shinyTextInputInfo,
-  shinyCheckboxInputInfo,
-  shinyCheckboxGroupInputInfo,
-  shinySelectInputInfo,
-  shinyRadioButtonsInfo,
-  shinyPlotOutputInfo,
-  shinyTextOutputInfo,
-  shinyUiOutputInfo,
-  shinyNavbarPageInfo,
-  shinyLayoutSidebarInfo,
-  shinyPanelMainInfo,
-  shinyTabsetPanelInfo,
-  gridlayoutGridPageInfo,
-  gridlayoutCardInfo,
-  gridlayoutTextPanelInfo,
-  gridlayoutGridCardPlotInfo,
-  gridlayoutGridContainerInfo,
-  dtDTOutputInfo,
-  bslibCardInfo,
-  bslibValueBoxInfo,
-  bslibCardBodyInfo,
-  bslibCardHeaderInfo,
-  bslibCardFooterInfo,
-  bslibSidebar,
-  bslibNavPanelInfo,
-  plotlyPlotlyOutputInfo,
-  textNodeInfo,
-  markdownNodeInfo,
-  unknownUiFunctionInfo,
-  testingErrorNodeInfo,
-] as const;
+import { registered_ui_nodes } from "./uiNodeTypes";
 
 type ShinyUiNodeInfo = (typeof registered_ui_nodes)[number];
 const shinyUiNodeInfo = new Map<string, ShinyUiNodeInfo>(
